Restrict access based on login status
Create require_session method: redirects to login page if user is not
logged in, then redirects back after the user logs in.

Create forbid_session method: redirects to overview page if user is
already logged in.

Create methods to remember current location or referer. Also create
method to redirect back to the remembered location or a default
location.