Improve 'back buttons'
Simplify methods for remembering pages and returning to them. Get rid of
the redirect_back_or_default method. Instead, we can use `redirect_to
remembered_page || default`, which is more flexible and syntactically
similar.

Get rid of calls to remember_page on submit. It was being called every
time the page loaded, which is not what we want. Instead use the
remember_referer method. This can be dangerous if two pages both
remember their referer and link to each other (causes loops), but I
haven't found a good alternative yet.